refactor(plugins): fix plugin loading data flow
