+++ zellij-server/src/tab/mod.rs
@@ -23,15 +23,14 @@ use crate::{
     panes::sixel::SixelImageStore,
     panes::{FloatingPanes, TiledPanes},
     panes::{LinkHandler, PaneId, PluginPane, TerminalPane},
+    plugins::PluginInstruction,
     pty::{ClientOrTabIndex, PtyInstruction, VteBytes},
     thread_bus::ThreadSenders,
-    wasm_vm::PluginInstruction,
     ClientId, ServerInstruction,
 };
 use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
 use std::rc::Rc;
-use std::sync::mpsc::channel;
 use std::time::Instant;
 use std::{
     collections::{HashMap, HashSet},
@@ -41,7 +40,7 @@ use zellij_utils::{
     data::{Event, InputMode, ModeInfo, Palette, PaletteColor, Style},
     input::{
         command::TerminalAction,
-        layout::{PaneLayout, Run},
+        layout::{PaneLayout, Run, RunPluginLocation},
         parse_keys,
     },
     pane_size::{Offset, PaneGeom, Size, SizeInPixels, Viewport},
@@ -70,13 +69,18 @@ macro_rules! resize_pty {
     }};
 }
 
-type HoldForCommand = Option<RunCommand>;
-
 // FIXME: This should be replaced by `RESIZE_PERCENT` at some point
 pub const MIN_TERMINAL_HEIGHT: usize = 5;
 pub const MIN_TERMINAL_WIDTH: usize = 5;
 
 const MAX_PENDING_VTE_EVENTS: usize = 7000;
+
+type HoldForCommand = Option<RunCommand>;
+
+enum BufferedTabInstruction {
+    SetPaneSelectable(PaneId, bool),
+    HandlePtyBytes(u32, VteBytes),
+}
 
 pub(crate) struct Tab {
     pub index: usize,
@@ -113,8 +117,11 @@ pub(crate) struct Tab {
     terminal_emulator_color_codes: Rc<RefCell<HashMap<usize, String>>>,
     pids_waiting_resize: HashSet<u32>, // u32 is the terminal_id
     cursor_positions_and_shape: HashMap<ClientId, (usize, usize, String)>, // (x_position,
-                                       // y_position,
-                                       // cursor_shape_csi)
+    // y_position,
+    // cursor_shape_csi)
+    is_pending: bool, // a pending tab is one that is still being loaded or otherwise waiting
+    pending_instructions: Vec<BufferedTabInstruction>, // instructions that came while the tab was
+                      // pending and need to be re-applied
 }
 
 #[derive(Clone, Debug, Default, Serialize, Deserialize)]
@@ -489,13 +496,16 @@ impl Tab {
             terminal_emulator_color_codes,
             pids_waiting_resize: HashSet::new(),
             cursor_positions_and_shape: HashMap::new(),
+            is_pending: true, // will be switched to false once the layout is applied
+            pending_instructions: vec![],
         }
     }
 
     pub fn apply_layout(
         &mut self,
         layout: PaneLayout,
-        new_ids: Vec<(u32, HoldForCommand)>,
+        new_terminal_ids: Vec<(u32, HoldForCommand)>,
+        mut new_plugin_ids: HashMap<RunPluginLocation, Vec<u32>>,
         tab_index: usize,
         client_id: ClientId,
     ) -> Result<()> {
@@ -523,7 +533,7 @@ impl Tab {
         match layout.position_panes_in_space(&free_space) {
             Ok(positions_in_layout) => {
                 let positions_and_size = positions_in_layout.iter();
-                let mut new_ids = new_ids.iter();
+                let mut new_terminal_ids = new_terminal_ids.iter();
 
                 let mut focus_pane_id: Option<PaneId> = None;
                 let mut set_focus_pane_id = |layout: &PaneLayout, pane_id: PaneId| {
@@ -535,18 +545,16 @@ impl Tab {
                 for (layout, position_and_size) in positions_and_size {
                     // A plugin pane
                     if let Some(Run::Plugin(run)) = layout.run.clone() {
-                        let (pid_tx, pid_rx) = channel();
+                        // let (pid_tx, pid_rx) = channel();
                         let pane_title = run.location.to_string();
-                        self.senders
-                            .send_to_plugin(PluginInstruction::Load(
-                                pid_tx,
-                                run,
-                                tab_index,
-                                client_id,
-                                position_and_size.into(),
-                            ))
-                            .with_context(err_context)?;
-                        let pid = pid_rx.recv().with_context(err_context)?;
+                        let pid = new_plugin_ids
+                            .get_mut(&run.location)
+                            .unwrap()
+                            .pop()
+                            .unwrap(); // TODO:
+                                       // err_context
+                                       // and
+                                       // stuff
                         let mut new_plugin = PluginPane::new(
                             pid,
                             *position_and_size,
@@ -570,7 +578,7 @@ impl Tab {
                         set_focus_pane_id(layout, PaneId::Plugin(pid));
                     } else {
                         // there are still panes left to fill, use the pids we received in this method
-                        if let Some((pid, hold_for_command)) = new_ids.next() {
+                        if let Some((pid, hold_for_command)) = new_terminal_ids.next() {
                             let next_terminal_position = self.get_next_terminal_position();
                             let initial_title = match &layout.run {
                                 Some(Run::Command(run_command)) => Some(run_command.to_string()),
@@ -601,7 +609,7 @@ impl Tab {
                         }
                     }
                 }
-                for (unused_pid, _) in new_ids {
+                for (unused_pid, _) in new_terminal_ids {
                     // this is a bit of a hack and happens because we don't have any central location that
                     // can query the screen as to how many panes it needs to create a layout
                     // fixing this will require a bit of an architecture change
@@ -639,20 +647,38 @@ impl Tab {
                         },
                     }
                 }
+                self.is_pending = false;
+                self.apply_buffered_instructions()?;
                 Ok(())
             },
             Err(e) => {
-                for (unused_pid, _) in new_ids {
+                for (unused_pid, _) in new_terminal_ids {
                     self.senders
                         .send_to_pty(PtyInstruction::ClosePane(PaneId::Terminal(unused_pid)))
                         .with_context(err_context)?;
                 }
+                self.is_pending = false;
                 Err::<(), _>(anyError::msg(e))
                     .with_context(err_context)
                     .non_fatal(); // TODO: propagate this to the user
                 Ok(())
             },
         }
+    }
+    pub fn apply_buffered_instructions(&mut self) -> Result<()> {
+        let buffered_instructions: Vec<BufferedTabInstruction> =
+            self.pending_instructions.drain(..).collect();
+        for buffered_instruction in buffered_instructions {
+            match buffered_instruction {
+                BufferedTabInstruction::SetPaneSelectable(pane_id, selectable) => {
+                    self.set_pane_selectable(pane_id, selectable);
+                },
+                BufferedTabInstruction::HandlePtyBytes(terminal_id, bytes) => {
+                    self.handle_pty_bytes(terminal_id, bytes)?;
+                },
+            }
+        }
+        Ok(())
     }
     pub fn update_input_modes(&mut self) -> Result<()> {
         // this updates all plugins with the client's input mode
@@ -1120,6 +1146,11 @@ impl Tab {
                 .any(|s_p| s_p.pid() == PaneId::Plugin(plugin_id))
     }
     pub fn handle_pty_bytes(&mut self, pid: u32, bytes: VteBytes) -> Result<()> {
+        if self.is_pending {
+            self.pending_instructions
+                .push(BufferedTabInstruction::HandlePtyBytes(pid, bytes));
+            return Ok(());
+        }
         let err_context = || format!("failed to handle pty bytes from fd {pid}");
         if let Some(terminal_output) = self
             .tiled_panes
@@ -1818,6 +1849,11 @@ impl Tab {
             .collect()
     }
     pub fn set_pane_selectable(&mut self, id: PaneId, selectable: bool) {
+        if self.is_pending {
+            self.pending_instructions
+                .push(BufferedTabInstruction::SetPaneSelectable(id, selectable));
+            return;
+        }
         if let Some(pane) = self.tiled_panes.get_pane_mut(id) {
             pane.set_selectable(selectable);
             if !selectable {
@@ -2805,6 +2841,10 @@ impl Tab {
         if let Some(active_pane) = self.get_active_pane_or_floating_pane_mut(client_id) {
             active_pane.clear_search();
         }
+    }
+
+    pub fn is_pending(&self) -> bool {
+        self.is_pending
     }
 
     fn show_floating_panes(&mut self) {
